Previous submission checker file renamed
